+++ ColorsDescriptor.xcodeproj/xcuserdata/belal.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "0ACEA610-676A-4422-B0C7-A19BCCA043C9"
+   type = "1"
+   version = "2.0">
+</Bucket>
