+++ ColorsDescriptor/ToolBar/ColorsToolBar.swift
@@ -10,7 +10,11 @@ import UIKit
 
 
 class ColorsToolBar : UIToolbar {
+    
     @IBOutlet weak var toolbarLabel: UIBarButtonItem!
+    @IBOutlet weak var colorsToolBar: ColorsToolBar!
+    
+    let customIidentifier = "ColorsToolBar"
     
     override func awakeFromNib() {
         super.awakeFromNib()
@@ -18,7 +22,32 @@ class ColorsToolBar : UIToolbar {
     
     func setup(){
         self.backgroundColor = .blue
-//        self.addSubview(UIButton())
-        toolbarLabel.title = "Hello"
+        //        self.addSubview(UIButton())
+        //        toolbarLabel.title = "Hello"
+    }
+    
+//    override init(frame: CGRect) {
+//        super.init(frame: frame)
+//        initSubviews()
+//    }
+//    
+//    required init?(coder aDecoder: NSCoder) {
+//        super.init(coder: aDecoder)
+//        initSubviews()
+//    }
+    
+    func initSubviews() {
+        
+        let nib = UINib(nibName: customIidentifier, bundle: nil)
+        awakeFromNib()
+
+        guard let view = nib.instantiate(withOwner: self, options: nil).first as?
+                UIView else {fatalError("Unable to convert nib")}
+        
+        view.frame = bounds
+        view.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        
+        addSubview(view)
+        
     }
 }
